Move logic from DeleteButton to EventDestroyer
