+++ client/app/components/calendar/event/Event.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import EventSchema from 'schemas/EventSchema';
-import DeleteButton from './DeleteButton';
+import EventDestroyer from './EventDestroyer';
 import { formatTime } from 'helpers/DateHelper';
 import './event.scss';
 
@@ -23,9 +23,8 @@ export default class Event extends React.Component {
 
     return (
       <div className="event" style={this._eventStyle()}>
-        <DeleteButton id={event.id}
-                      onDelete={this.props.onDelete}
-                      disabled={!creator.self} />
+        <EventDestroyer onDelete={this.props.onDelete}
+                        disabled={!creator.self} />
         <div className="event-time">{timeStr}</div>
         <div className="event-name">{event.summary}</div>
         <div className="event-user">
