+++ client/test/components/calendar/event/EventDestroyer.spec.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import DeleteButton from 'components/calendar/event/DeleteButton';
+import EventDestroyer from 'components/calendar/event/EventDestroyer';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { Overlay, Modal } from 'react-bootstrap';
 
-describe('<DeleteButton />', () => {
+describe('<EventDestroyer />', () => {
   let spy;
   beforeEach(() => {
     spy = sinon.spy();
   });
 
   it('renders', () => {
-    const wrapper = mount(<DeleteButton onDelete={spy} disabled={false} />);
+    const wrapper = mount(<EventDestroyer onDelete={spy} disabled={false} />);
     expect(wrapper.find('span').props().className).to.contain('delete-button');
     expect(wrapper.find('span').props().className).to.contain('glyphicon');
   });
@@ -21,7 +21,7 @@ describe('<DeleteButton />', () => {
     context('props.disabled', () => {
       let wrapper;
       beforeEach(() => {
-        wrapper = mount(<DeleteButton onDelete={spy} disabled />);
+        wrapper = mount(<EventDestroyer onDelete={spy} disabled />);
       });
 
       it('has information overlay', () => {
@@ -33,7 +33,7 @@ describe('<DeleteButton />', () => {
     context('props.disabled === false', () => {
       let wrapper;
       beforeEach(() => {
-        wrapper = mount(<DeleteButton onDelete={spy} disabled={false} />);
+        wrapper = mount(<EventDestroyer onDelete={spy} disabled={false} />);
       });
 
       it('has a confirmation box', () => {
